adding the MIT license explicitly to the source
I have to add the license explicitly to the source, because some people, just don't understand how licensing works and show no respect in regard to open-source licensed and copyright. The package is marked as MIT licensed. The license is in the root and applies to all files. Even if you unfriendly fork, you should never change the MIT license.

And i really don't like, what i see over here: 

Lawondyss/CasperJS-Installer

First the repository is git forked. Then the source is modified, but badly, leaving some traces to the original work behind.
The license file in the root is removed and the license annotation inside the composer.json file is modified to BSD.
The source is now committed as a new Git repository. Then the BSD license is changed to BSD-2-clause. 
Seriously?

I mean, thank you for forking - that means that you like my installer, which is based on two really dirts hacks.
The first one is the creation of an composer in-memory package.
The second one is re- or misusing the outer package version as the inner version.
Both originate from this installer. 

And: it's not only a violation of the MIT license - it's also a copyright violation, which affects me and all 
the other contributors (https://github.com/jakoch/phantomjs-installer/graphs/contributors), which helped to improve the installer over time.